+++ scrm-service/src/main/java/com/huotu/scrm/service/service/ReportDayService.java
@@ -12,5 +12,5 @@ public interface ReportDayService {
      *
      * @param userId 用户ID
      */
-    ApiResult saveReportDay(int userId);
+    ApiResult saveReportDay(long userId);
 }
